+++ core/python_expression_dsl/parser.py
@@ -1,12 +1,10 @@
 import re
-from typing import Optional, Type, TypeVar
 
 from pyparsing import (
     Forward,
     Group,
     Literal,
     ParseException,
-    ParseResults,
     QuotedString,
     Regex,
     Suppress,
@@ -16,24 +14,21 @@ from pyparsing import (
     alphas,
 )
 
-from ..exceptions import BaseError
-from ..util import chunks
-from .ast import (
-    BinaryArithmeticExpression,
-    BinaryBooleanExpression,
-    BinaryExpression,
-    ComparisonExpression,
-    DotExpression,
-    Expression,
-    FunctionCallExpression,
-    Identifier,
-    Number,
-    Operator,
-    SliceExpression,
-    String,
-    UnaryArithmeticExpression,
-    UnaryBooleanExpression,
-    UnaryExpression,
+from core.exceptions import BaseError
+from core.python_expression_dsl.ast import Operator
+from core.python_expression_dsl.util import (
+    _parse_binary_arithmetic_expression,
+    _parse_binary_boolean_expression,
+    _parse_comparison_expression,
+    _parse_dot_expression,
+    _parse_function_call_expression,
+    _parse_identifier,
+    _parse_number,
+    _parse_parenthesized_expression,
+    _parse_slice_operation,
+    _parse_string,
+    _parse_unary_arithmetic_expression,
+    _parse_unary_boolean_expression,
 )
 
 
@@ -48,196 +43,6 @@ class DSLParser(object):
         r"Expected\s+(\{.+\}),\s+found\s+('.+')\s+\(at\s+char\s+(\d+)\)"
     )
     DEFAULT_ERROR_MESSAGE = "Could not parse the expression"
-
-    @staticmethod
-    def _parse_identifier(tokens: ParseResults) -> Identifier:
-        """Transform the token into an Identifier node.
-
-        :param tokens: ParseResults objects
-
-        :return: Identifier node
-        """
-        return Identifier(tokens[0])
-
-    @staticmethod
-    def _parse_string(tokens: ParseResults) -> String:
-        """Transform the token into a String node.
-
-        :param tokens: ParseResults objects
-
-        :return: Identifier node
-        """
-        return String(tokens[0])
-
-    @staticmethod
-    def _parse_number(tokens: ParseResults) -> Number:
-        """Transform the token into a Number node.
-
-        :param tokens: ParseResults objects
-
-        :return: Number node
-        """
-        return Number(tokens[0])
-
-    UE = TypeVar("UE", bound=UnaryExpression)
-
-    @staticmethod
-    def _parse_unary_expression(
-        expression_type: Type[UE], tokens: ParseResults
-    ) -> Optional[UE]:
-
-        """Transform the token into an unary expression.
-
-        :param tokens: ParseResults objects
-
-        :return: UnaryExpression node
-        """
-        if len(tokens) >= 2:
-            tokens = list(reversed(tokens))
-            argument = tokens[0]
-            operator_type = tokens[1]
-            expression = expression_type(operator_type, argument)
-
-            for tokens_chunk in chunks(tokens, 1, 2):
-                operator_type = tokens_chunk[0]
-                expression = expression_type(operator_type, expression)
-
-            return expression
-        else:
-            return None
-
-    @staticmethod
-    def _parse_unary_arithmetic_expression(
-        tokens: ParseResults,
-    ) -> Optional[UnaryArithmeticExpression]:
-        """Transform the token into an UnaryArithmeticExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: UnaryArithmeticExpression node
-        """
-        return DSLParser._parse_unary_expression(UnaryArithmeticExpression, tokens)
-
-    @staticmethod
-    def _parse_unary_boolean_expression(
-        tokens: ParseResults,
-    ) -> Optional[UnaryBooleanExpression]:
-        """Transform the token into an UnaryBooleanExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: UnaryBooleanExpression node
-        """
-        return DSLParser._parse_unary_expression(UnaryBooleanExpression, tokens)
-
-    BE = TypeVar("BE", bound=BinaryExpression)
-
-    @staticmethod
-    def _parse_binary_expression(
-        expression_type: Type[BE], tokens: ParseResults
-    ) -> Optional[BE]:
-        """Transform the token into a BinaryExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: BinaryExpression node
-        """
-        if len(tokens) >= 3:
-            left_argument = tokens[0]
-            operator_type = tokens[1]
-            right_argument = tokens[2]
-            expression = expression_type(operator_type, left_argument, right_argument)
-
-            for tokens_chunk in chunks(tokens, 2, 3):
-                operator_type = tokens_chunk[0]
-                right_argument = tokens_chunk[1]
-                expression = expression_type(operator_type, expression, right_argument)
-
-            return expression
-        else:
-            return None
-
-    @staticmethod
-    def _parse_binary_arithmetic_expression(
-        tokens: ParseResults,
-    ) -> Optional[BinaryArithmeticExpression]:
-        """Transform the token into a BinaryArithmeticExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: BinaryArithmeticExpression node
-        """
-        return DSLParser._parse_binary_expression(BinaryArithmeticExpression, tokens)
-
-    @staticmethod
-    def _parse_binary_boolean_expression(
-        tokens: ParseResults,
-    ) -> Optional[BinaryBooleanExpression]:
-        """Transform the token into a BinaryBooleanExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: BinaryBooleanExpression node
-        """
-        return DSLParser._parse_binary_expression(BinaryBooleanExpression, tokens)
-
-    @staticmethod
-    def _parse_comparison_expression(
-        tokens: ParseResults,
-    ) -> Optional[ComparisonExpression]:
-        """Transform the token into a ComparisonExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: ComparisonExpression node
-        """
-        return DSLParser._parse_binary_expression(ComparisonExpression, tokens)
-
-    @staticmethod
-    def _parse_dot_expression(tokens: ParseResults) -> DotExpression:
-        """Transform the token into a DotExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: ComparisonExpression node
-        """
-        return DotExpression(list(tokens[0]))
-
-    @staticmethod
-    def _parse_parenthesized_expression(tokens: ParseResults) -> Expression:
-        """Transform the token into a Expression node.
-
-        :param tokens: ParseResults objects
-
-        :return: ComparisonExpression node
-        """
-        return tokens[0]
-
-    @staticmethod
-    def _parse_function_call_expression(tokens: ParseResults) -> FunctionCallExpression:
-        """Transform the token into a FunctionCallExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: ComparisonExpression node
-        """
-        function_identifier = tokens[0][0]
-        arguments = tokens[0][1:]
-
-        return FunctionCallExpression(function_identifier, arguments)
-
-    @staticmethod
-    def _parse_slice_operation(tokens: ParseResults) -> SliceExpression:
-        """Transform the token into a SliceExpression node.
-
-        :param tokens: ParseResults objects
-
-        :return: SliceExpression node
-        """
-        array_expression = tokens[0][0]
-        slice_expression = tokens[0][1]
-
-        return SliceExpression(array_expression, slice_expression)
 
     # Auxiliary tokens
     LEFT_PAREN, RIGHT_PAREN = map(Suppress, "()")
@@ -285,15 +90,9 @@ class DSLParser(object):
         | IN_OPERATOR
     )
 
-    NUMBER = Regex(r"[+-]?\d+(?:\.\d*)?(?:[eE][+-]?\d+)?").setParseAction(
-        _parse_number.__func__  # type: ignore
-    )
-    IDENTIFIER = Word(alphas, alphanums + "_$").setParseAction(
-        _parse_identifier.__func__  # type: ignore
-    )
-    STRING = (QuotedString("'") | QuotedString('"')).setParseAction(
-        _parse_string.__func__  # type: ignore
-    )
+    NUMBER = Regex(r"[+-]?\d+(?:\.\d*)?(?:[eE][+-]?\d+)?").setParseAction(_parse_number)
+    IDENTIFIER = Word(alphas, alphanums + "_$").setParseAction(_parse_identifier)
+    STRING = (QuotedString("'") | QuotedString('"')).setParseAction(_parse_string)
 
     # Unary boolean operator
     INVERSION_OPERATOR = Literal("not").setParseAction(lambda _: Operator.INVERSION)
@@ -306,54 +105,38 @@ class DSLParser(object):
 
     comparison_expression = (
         arithmetic_expression + ZeroOrMore(COMPARISON_OPERATOR + arithmetic_expression)
-    ).setParseAction(
-        _parse_comparison_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_comparison_expression)
 
     inversion_expression = (
         ZeroOrMore(INVERSION_OPERATOR) + comparison_expression
-    ).setParseAction(
-        _parse_unary_boolean_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_unary_boolean_expression)
     conjunction_expression = (
         inversion_expression + ZeroOrMore(CONJUNCTION_OPERATOR + inversion_expression)
-    ).setParseAction(
-        _parse_binary_boolean_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_binary_boolean_expression)
     disjunction_expression = (
         conjunction_expression
         + ZeroOrMore(DISJUNCTION_OPERATOR + conjunction_expression)
-    ).setParseAction(
-        _parse_binary_boolean_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_binary_boolean_expression)
 
     expression = disjunction_expression
 
     dot_expression = Group(
         IDENTIFIER + ZeroOrMore(FULL_STOP + expression)
-    ).setParseAction(
-        _parse_dot_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_dot_expression)
 
     parenthesized_expression = Group(
         LEFT_PAREN + expression + RIGHT_PAREN
-    ).setParseAction(
-        _parse_parenthesized_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_parenthesized_expression)
 
     slice = expression
     slice_expression = Group(
         IDENTIFIER + LEFT_BRACKET + slice + RIGHT_BRACKET
-    ).setParseAction(
-        _parse_slice_operation.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_slice_operation)
 
     function_call_arguments = ZeroOrMore(expression + ZeroOrMore(COMMA + expression))
     function_call_expression = Group(
         IDENTIFIER + LEFT_PAREN + function_call_arguments + RIGHT_PAREN
-    ).setParseAction(
-        _parse_function_call_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_function_call_expression)
 
     atom = (
         ZeroOrMore(NEGATION_OPERATOR)
@@ -366,22 +149,18 @@ class DSLParser(object):
             | dot_expression
             | IDENTIFIER
         )
-    ).setParseAction(
-        _parse_unary_arithmetic_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_unary_arithmetic_expression)
 
     factor = Forward()
     factor << (atom + ZeroOrMore(POWER_OPERATOR + factor)).setParseAction(
-        _parse_binary_arithmetic_expression.__func__  # type: ignore
+        _parse_binary_arithmetic_expression
     )
     term = (factor + ZeroOrMore(MULTIPLICATIVE_OPERATOR + factor)).setParseAction(
-        _parse_binary_arithmetic_expression.__func__  # type: ignore
+        _parse_binary_arithmetic_expression
     )
     arithmetic_expression << (
         term + ZeroOrMore(ADDITIVE_OPERATOR + term)
-    ).setParseAction(
-        _parse_binary_arithmetic_expression.__func__
-    )  # type: ignore
+    ).setParseAction(_parse_binary_arithmetic_expression)
 
     def _parse_error_message(self, parse_exception):
         """Transform the standard error description into a readable concise message.
